Change manager priority attribute to float.
This used to be revision r2507.

+++ core/manager/src/main/java/imagej/manager/Manager.java
@@ -51,6 +51,6 @@ import net.java.sezpoz.Indexable;
 @Indexable(type=ManagerComponent.class)
 public @interface Manager {
 
-	int priority() default Managers.NORMAL_PRIORITY;
+	float priority() default Managers.NORMAL_PRIORITY;
 
 }
